coercions for systems (lije Solaris) on which isspace etc are macros
git-svn-id: https://svn.r-project.org/R/branches/R-1-2-patches@11953 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/main/util.c
@@ -657,7 +657,7 @@ SEXP type2str(SEXPTYPE t)
 Rboolean isBlankString(unsigned char *s)
 {
     while (*s)
-	if (!isspace(*s++)) return FALSE;
+	if (!isspace((int)*s++)) return FALSE;
     return TRUE;
 }
